kafka/producer_ccxt/requirements.txt
ccxt==4.0.85
kafka-python==2.0.2
psycopg2-binary